core-app-api: deprecate dependency on core-components
