+++ .changeset/giant-drinks-wave.md
@@ -1,0 +1,5 @@
+---
+'@backstage/dev-utils': patch
+---
+
+Migrated to explicit passing of components to `createApp`.
