+++ .changeset/late-rice-sit.md
@@ -1,0 +1,5 @@
+---
+'@backstage/core-components': patch
+---
+
+Added a new `defaultAppComponents` method that creates a minimal set of components to pass on to `createApp` from `@backstage/core-app-api`.
