+++ packages/core-components/api-report.md
@@ -6,6 +6,7 @@
 /// <reference types="react" />
 
 import { ApiRef } from '@backstage/core-plugin-api';
+import { AppComponents } from '@backstage/core-plugin-api';
 import { BackstageIdentityApi } from '@backstage/core-plugin-api';
 import { BackstagePalette } from '@backstage/theme';
 import { BackstageTheme } from '@backstage/theme';
@@ -198,6 +199,9 @@ export type CustomProviderClassKey = 'form' | 'button';
 
 // @public (undocumented)
 export function DashboardIcon(props: IconComponentProps): JSX.Element;
+
+// @public
+export function defaultAppComponents(): Omit<AppComponents, 'Router'>;
 
 // @public
 type DependencyEdge<T = {}> = T & {
