+++ packages/core-components/src/index.ts
@@ -25,3 +25,4 @@ export * from './hooks';
 export * from './icons';
 export * from './layout';
 export * from './overridableComponents';
+export { defaultAppComponents } from './defaultAppComponents';
